--- _19_Medicine/README.md
@@ -1,7 +1,0 @@
---- Day 19: Medicine for Rudolph ---
-===
-
-
-Tests
----
-        go test
